update dan menambahkan fitur perpustakaan

+++ database/migrations/2026_03_31_065859_add_role_to_users_table.php
@@ -9,14 +9,18 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->enum('role', ['anggota', 'petugas', 'kepala'])->default('anggota')->after('email');
+            if (!Schema::hasColumn('users', 'role')) {
+                $table->enum('role', ['anggota', 'petugas', 'kepala'])->default('anggota')->after('email');
+            }
         });
     }
 
     public function down(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('role');
+            if (Schema::hasColumn('users', 'role')) {
+                $table->dropColumn('role');
+            }
         });
     }
 };
